requirements.txt
tensorflow
pandas 
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
python-dotenv
-e .
streamlit
